requirements.txt
Package         Version
--------------- -------
aniso8601       9.0.1
attrs           23.1.0
blinker         1.6.2
click           8.1.3
flasgger        0.9.5
Flask           2.3.2
Flask-RESTful   0.3.9
itsdangerous    2.1.2
Jinja2          3.1.2
jsonschema      4.17.3
MarkupSafe      2.1.2
marshmallow     3.19.0
mistune         2.0.5
packaging       23.1
pip             22.3.1
psycopg2-binary 2.9.6
pyrsistent      0.19.3
pytz            2023.3
PyYAML          6.0
setuptools      65.5.1
six             1.16.0
Werkzeug        2.3.3
wheel           0.38.4
